requirements.txt
hidapi
paho-mqtt
